+++ src/shared/components/case-editor/case-edit-page/case-edit-page.component.ts
@@ -15,11 +15,11 @@ import { FormValueService } from '../../../services/form/form-value.service';
 import { SaveOrDiscardDialogComponent } from '../../dialogs/save-or-discard-dialog';
 import { CallbackErrorsContext } from '../../error/domain/error-context';
 import { initDialog } from '../../helpers';
+import { LinkedCasesError } from '../../palette';
 import { CaseEditComponent } from '../case-edit/case-edit.component';
 import { WizardPage } from '../domain/wizard-page.model';
 import { Wizard } from '../domain/wizard.model';
 import { PageValidationService } from '../services/page-validation.service';
-import { LinkedCasesError } from '../../palette';
 
 @Component({
   selector: 'ccd-case-edit-page',