+++ src/shared/components/case-viewer/case-event-trigger/case-event-trigger.component.spec.ts
@@ -232,6 +232,14 @@ describe('CaseEventTriggerComponent', () => {
     expect(router.navigate).toHaveBeenCalledWith(['/' + URL_SEGMENTS[0].path + '/' + URL_SEGMENTS[1].path]);
   });
 
+  it('should cancel navigate to linked cases tab', () => {
+    const routerWithModifiedUrl = TestBed.get(Router);
+    routerWithModifiedUrl.url = 'linkCases';
+    component.caseDetails.case_id = '1111-2222-3333-4444';
+    component.cancel();
+    expect(router.navigate).toHaveBeenCalledWith(['cases', 'case-details', '1111-2222-3333-4444']);
+  });
+
   it('should bypass validation if the CaseEventData data object contains a FlagLauncher field', (done) => {
     CASE_DETAILS.tabs = [
       {