+++ src/shared/components/palette/case-link/components/link-cases/link-cases.component.ts
@@ -31,7 +31,6 @@ export class LinkCasesComponent implements OnInit {
   public caseReasonError: string;
   public caseSelectionError: string;
   public noSelectedCaseError: string;
-  public caseName = 'Case name missing';
   private ISO_FORMAT = 'YYYY-MM-DDTHH:mm:ss.SSS';
 
   constructor(
@@ -165,7 +164,7 @@ export class LinkCasesComponent implements OnInit {
             caseType: this.linkedCasesService.mapLookupIDToValueFromJurisdictions('CASE_TYPE', caseView.case_type.id),
             caseState: this.linkedCasesService.mapLookupIDToValueFromJurisdictions('STATE', caseView.state.name),
             caseService: this.linkedCasesService.mapLookupIDToValueFromJurisdictions('JURISDICTION', caseView.case_type.jurisdiction.name),
-            caseName: caseView.metadataFields && caseView.metadataFields['caseNameHmctsInternal'] || 'Case name missing',
+            caseName: this.linkedCasesService.getCaseName(caseView),
           };
           const ccdApiCaseLinkData: CCDCaseLinkType = {
             CaseReference: caseView.case_id,